Image showing how to place the newer LED-buttons without the plastic pin guide.
Image showing how to place the newer LED-buttons without the plastic pin guide. 
The new component kits contains a new kind of LED-button with a lower height. It does not have the plastic "pin" used to place it correctly. Instead it has a small dot of the color of the LED. 
It is very important that this dot is place AWAY from the closets edge of the PCB. 
Another way of saying it is that the dot should face the potentiometers. If this does not help, please contact me.